Address comment from PR
