+++ bundles/org.eclipse.osgi/container/src/org/eclipse/osgi/internal/serviceregistry/ServiceRegistrationImpl.java
@@ -19,6 +19,7 @@ import java.util.Dictionary;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.ReentrantLock;
 import org.eclipse.osgi.internal.debug.Debug;
 import org.eclipse.osgi.internal.framework.BundleContextImpl;
 import org.eclipse.osgi.internal.loader.sources.PackageSource;
@@ -537,42 +538,49 @@ public class ServiceRegistrationImpl<S> implements ServiceRegistration<S>, Compa
 			}
 
 			/* Obtain and return the service object */
-			ReentrantLockExt useLock = use.getLock();
+			ReentrantLock useLock = use.getLock();
 			try {
-				if (useLock.tryLock(use.getLockTimeout(), TimeUnit.SECONDS)) {
-					if (registry.debug.DEBUG_SERVICES) {
-						Debug.println('[' + Thread.currentThread().getName() + "] getServiceLock[" //$NON-NLS-1$
-								+ user.getBundleImpl() + "](" + this + "), id:" + System.identityHashCode(useLock) //$NON-NLS-1$ //$NON-NLS-2$
-								+ ". holdCount:" + useLock.getHoldCount() //$NON-NLS-1$
-								+ ", queued:" + useLock.getQueueLength()); //$NON-NLS-1$
-					}
+				if (useLock.tryLock(use.getLockTimeout().toMillis(), TimeUnit.MILLISECONDS)) {
+					try {
+						if (registry.debug.DEBUG_SERVICES) {
+							Debug.println('[' + Thread.currentThread().getName() + "] getServiceLock[" //$NON-NLS-1$
+									+ user.getBundleImpl() + "](" + this + "), id:" + System.identityHashCode(useLock) //$NON-NLS-1$ //$NON-NLS-2$
+									+ ". holdCount:" + useLock.getHoldCount() //$NON-NLS-1$
+									+ ", queued:" + useLock.getQueueLength()); //$NON-NLS-1$
+						}
 
-					/* if another thread removed the ServiceUse, then
-					 * go back to the top and start again */
-					synchronized (servicesInUse) {
-						user.checkValid();
-						if (servicesInUse.get(this) != use) {
-							if (registry.debug.DEBUG_SERVICES) {
-								Debug.println('[' + Thread.currentThread().getName() + "] getServiceContinue[" //$NON-NLS-1$
-										+ user.getBundleImpl() + "](" + this //$NON-NLS-1$
-										+ ")"); //$NON-NLS-1$
-							}
-							continue;
-						}
-					}
-					S serviceObject = consumer.getService(use);
-					/* if the service factory failed to return an object and
-					 * we created the service use, then remove the 
-					 * optimistically added ServiceUse. */
-					if ((serviceObject == null) && added) {
+						/*
+						 * if another thread removed the ServiceUse, then go back to the top and start
+						 * again
+						 */
 						synchronized (servicesInUse) {
-							synchronized (registrationLock) {
-								servicesInUse.remove(this);
-								contextsUsing.remove(user);
+							user.checkValid();
+							if (servicesInUse.get(this) != use) {
+								if (registry.debug.DEBUG_SERVICES) {
+									Debug.println('[' + Thread.currentThread().getName() + "] getServiceContinue[" //$NON-NLS-1$
+											+ user.getBundleImpl() + "](" + this //$NON-NLS-1$
+											+ ")"); //$NON-NLS-1$
+								}
+								continue;
 							}
 						}
+						S serviceObject = consumer.getService(use);
+						/*
+						 * if the service factory failed to return an object and we created the service
+						 * use, then remove the optimistically added ServiceUse.
+						 */
+						if ((serviceObject == null) && added) {
+							synchronized (servicesInUse) {
+								synchronized (registrationLock) {
+									servicesInUse.remove(this);
+									contextsUsing.remove(user);
+								}
+							}
+						}
+						return serviceObject;
+					} finally {
+						useLock.unlock();
 					}
-					return serviceObject;
 				}
 				throw new ServiceException(
 						"Failed to acquire lock within try period, Lock, id:" + System.identityHashCode(useLock) + ", " //$NON-NLS-1$ //$NON-NLS-2$
@@ -581,10 +589,6 @@ public class ServiceRegistrationImpl<S> implements ServiceRegistration<S>, Compa
 			} catch (InterruptedException e) {
 				Thread.currentThread().interrupt(); // To appease sonar
 				return null;
-			} finally {
-				if (useLock.isHeldByCurrentThread()) {
-					useLock.unlock();
-				}
 			}
 		}
 	}
@@ -659,7 +663,7 @@ public class ServiceRegistrationImpl<S> implements ServiceRegistration<S>, Compa
 
 		return use.callWithLock(() -> {
 			if (registry.debug.DEBUG_SERVICES) {
-				ReentrantLockExt useLock = use.getLock();
+				ReentrantLock useLock = use.getLock();
 				Debug.println('[' + Thread.currentThread().getName() + "] ungetServiceLock[" + user.getBundleImpl() //$NON-NLS-1$
 						+ "](" + this + "), id:" + System.identityHashCode(useLock) + ", holdCount:" //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
 						+ useLock.getHoldCount()
